optimised TCP port scanner
The transmission control protocol is a system that ensures secure communication between the client and server. And for the sake of security and efficiency, it is necessary to determine which of the 65,535 ports are available. This code block is a simple and yet optimised way using go-routines and channels by which ports on the computer can be scanned to see which are closed and which are open.